Logging Additional Data
This is the first of a series of updates which will increase the amount of data that is logged regarding the process that is flagged.

I added additional logic to this code to pull information about the flagged process. (Natively, YARA does not provide much information other than the rule and the process that matches the rule.)